Fix user_id column type to bigint for MySQL compatibility

+++ db/schema.rb
@@ -16,7 +16,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_05_02_055102) do
     t.string "ip_address"
     t.datetime "updated_at", null: false
     t.string "user_agent"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.index ["user_id"], name: "index_sessions_on_user_id"
   end
 
@@ -27,7 +27,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_05_02_055102) do
     t.integer "difficulty"
     t.integer "priority"
     t.datetime "updated_at", null: false
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.index ["user_id"], name: "index_tasks_on_user_id"
   end
